Input to the normal network is now shaped right.
Now it flattens the feature maps into one row
of a matrix with each row representing a single item in the
minibatch at the end of the convolution and maxpooling layers.
This matrix then gets passed as the input to the normal part of
network, with is how it expects it.